Move APIs to be used by both controller and configurer
